Move canopy_light_helpers out of core
