+++ src/module_library/AuxBioCro.cpp
@@ -10,7 +10,7 @@
 #include "c4photo.h"
 #include "BioCro.h"
 #include "boundary_layer_conductance.h"  // for leaf_boundary_layer_conductance_nikolov
-#include "core/canopy_light_helpers.h"   // for core::thick_layer_absorption
+#include "canopy_light_helpers.h"        // for core::thick_layer_absorption
 #include "water_and_air_properties.h"    // for saturation_vapor_pressure,
                                          // TempToDdryA, TempToLHV, TempToSFS
 #include "../framework/constants.h"      // for pi, e, molar_mass_of_water,
