+++ src/module_library/canopy_light_helpers.cpp
